refactor: move packages to folders
remove unused code
