+++ src/packages/commit-lint/config-conventional.package.ts
