--- src/packages/extensions/index.ts
@@ -1,3 +1,0 @@
-export * from './config-conventional.package';
-export * from './ts-jest.package';
-export * from './typs-jest.package';
