+++ src/packages/jest/typs-jest.package.ts
@@ -1,6 +1,9 @@
 import { DependencyTypeEnum, PackagesEnumKeys } from '../../type/enums';
 import { BasePackage } from '../base.package';
 
+/**
+ * @see https://www.npmjs.com/package/@types/jest
+ * */
 class TypesJest extends BasePackage {
   readonly dependencyType: DependencyTypeEnum =
     DependencyTypeEnum.devDependency;
