+++ src/packages/prettier/prettier.constant.ts
@@ -1,0 +1,22 @@
+export const CONFIG: string = JSON.stringify(
+  {
+    tabWidth: 2,
+    useTabs: false,
+    singleQuote: true,
+    semi: true,
+    bracketSpacing: true,
+    arrowParens: 'avoid',
+    trailingComma: 'es5',
+    bracketSameLine: true,
+    printWidth: 80,
+  },
+  undefined,
+  2
+);
+
+export const IGNORE: string = `# Add files here to ignore them from prettier formatting
+/dist
+/coverage
+.angular
+node_modules
+`;
