+++ src/services/node/child-process.service.ts
@@ -1,28 +1,8 @@
 import * as cp from 'node:child_process';
 
 class ChildProcessService {
-  exec(command: string): void {
-    cp.exec(command, this.#callback);
-  }
-
   execSync(command: string, options?: cp.ExecSyncOptions): string {
     return cp.execSync(command, { encoding: 'utf8', ...options }).toString();
-  }
-
-  #callback(
-    error: cp.ExecException | null,
-    stdout: string,
-    stderr: string
-  ): void {
-    if (error) {
-      console.info(`error: ${error.message}`);
-      return;
-    }
-    if (stderr) {
-      console.info(`stderr: ${stderr}`);
-      return;
-    }
-    console.info(stdout);
   }
 }
